backend/requirements.txt
email-validator==2.2.0
pydantic-settings==2.6.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
pypdf==5.1.0
python-docx==1.1.2
httpx==0.28.1
openai==1.57.4
sentence-transformers==2.7.0
